pyproject.toml
[project]
name = "tropical deepsea neutrino telescope (TRIDENT) machine learning project"
version = "0.1.0"
description = "A PyTorch-based regression framework for physics data using DeepSet and MLP architectures."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Kunz Bryan" }]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "matplotlib>=3.6.0",
    "pyarrow>=10.0.0",
    "PyYAML>=6.0.2",
    "dacite>=1.9.2",
    "simple-parsing>=0.1.7",
    "pydantic>=1.10.0",
    "tqdm>=4.64.0",
    "joblib"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["MLP_Data_Network"]